Optimize logic
