+++ vllm/lora/utils.py
@@ -13,7 +13,7 @@ from transformers import PretrainedConfig
 from vllm.config import LoRAConfig
 from vllm.logger import init_logger
 from vllm.lora.fully_sharded_layers import (
-    ColumnParallelLinearWithShardedLoRA, HFCompatibleLinearWithShardedLoRA,
+    ColumnParallelLinearWithShardedLoRA,
     MergedColumnParallelLinearWithShardedLoRA,
     MergedQKVParallelLinearWithShardedLora, QKVParallelLinearWithShardedLora,
     RowParallelLinearWithShardedLoRA)
@@ -21,7 +21,6 @@ from vllm.lora.fully_sharded_layers import (
 # yapf conflicts with isort for this block
 # yapf: disable
 from vllm.lora.layers import (BaseLayerWithLoRA, ColumnParallelLinearWithLoRA,
-                              HFCompatibleLinearWithLoRA,
                               LinearScalingRotaryEmbeddingWithLora,
                               LogitsProcessorWithLoRA,
                               MergedColumnParallelLinearWithLoRA,
@@ -46,14 +45,12 @@ _all_lora_classes: Set[Type[BaseLayerWithLoRA]] = {
     MergedQKVParallelLinearWithLora,
     RowParallelLinearWithLoRA,
     ReplicatedLinearWithLoRA,
-    HFCompatibleLinearWithLoRA,
     LogitsProcessorWithLoRA,
     ColumnParallelLinearWithShardedLoRA,
     QKVParallelLinearWithShardedLora,
     MergedColumnParallelLinearWithShardedLoRA,
     MergedQKVParallelLinearWithShardedLora,
     RowParallelLinearWithShardedLoRA,
-    HFCompatibleLinearWithShardedLoRA,
     LinearScalingRotaryEmbeddingWithLora,
 }
 
@@ -69,9 +66,20 @@ def from_layer(layer: nn.Module,
                                       lora_config=lora_config,
                                       packed_modules_list=packed_modules_list,
                                       model_config=model_config):
-            ret = lora_cls(layer)
-            ret.create_lora_weights(max_loras, lora_config, model_config)
-            return ret
+            instance_layer = lora_cls(layer)
+            if layer.__class__.__name__ == "HFCompatibleLinear":
+                # HACK:  Make the forward method compatible with the original
+                # forward method of the instance_layer.
+                original_forward = instance_layer.forward
+
+                def new_forward(input):
+                    input = input.squeeze(0)
+                    return original_forward(input)[0]  # noqa: B023
+
+                instance_layer.forward = new_forward
+            instance_layer.create_lora_weights(max_loras, lora_config,
+                                               model_config)
+            return instance_layer
 
 
 def from_layer_logits_processor(
